requirements.txt
fastapi
python-dotenv
openai